+++ avalonia/Artemis.Core/Models/CriteriaNode.cs
@@ -12,4 +12,4 @@ public record GroupNode(int Id, OperatorType Operator) : CriteriaNode(Id)
 {
     public OperatorType[] OperatorTypeValues { get; } = Enum.GetValues<OperatorType>();
 }
-public record TermNode(int Id, int CategoryId, decimal DistAmt) : CriteriaNode(Id);
+public record TermNode(int Id, int CategoryId, double DistAmt) : CriteriaNode(Id);
