+++ avalonia/Artemis.Core/Models/Location.cs
@@ -5,7 +5,7 @@ public record Location
     public int Id { get; set; }
     public string? Name { get; set; }
     public string? Address { get; set; }
-    public decimal? Latitude { get; set; }
-    public decimal? Longitude { get; set; }
+    public double? Latitude { get; set; }
+    public double? Longitude { get; set; }
     public string? Notes { get; set; }
 }
